backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
cryptography
pyjwt
passlib[bcrypt]
sqlalchemy
aiosqlite
